Refactor reducers into individual files and move contact specific redux stuff into Contact
