+++ client/src/redux/reducers.jsx
@@ -1,131 +1,10 @@
-import _find from 'lodash/find';
-import _merge from 'lodash/merge';
+import { combineReducers } from 'redux';
+import contact from './reducers/contact';
+import pageTitle from './reducers/pageTitle';
+import projects from './reducers/projects';
 
-import { SET_TITLE, FETCH_PROJECTS, RECEIVE_PROJECTS, SEND_EMAIL, RECEIVE_EMAIL, DISMISS_SUCCESS,
-  DISMISS_ERROR, UPDATE_FIELD, UPDATE_FIELDS } from './actions';
-
-const initialState = {
-  pageTitle: 'Page Title',
-  projects: {
-    list: [],
-    fetching: false
-  },
-  contact: {
-    isLoading: false,
-    showSuccess: false,
-    showError: false,
-    fields: [
-      {
-        name: 'name',
-        value: '',
-        valid: true,
-        errorMsg: ''
-      },
-      {
-        name: 'email',
-        value: '',
-        valid: true,
-        errorMsg: ''
-      },
-      {
-        name: 'message',
-        value: '',
-        valid: true,
-        errorMsg: ''
-      },
-      {
-        name: 'captcha',
-        value: '',
-        valid: true,
-        errorMsg: ''
-      }
-    ]
-  }
-};
-
-function reduce(state = initialState, action) {
-  switch (action.type) {
-    case SET_TITLE:
-      document.title = `${action.text} - Michael Humiston`;
-      return _merge({}, state, {
-        pageTitle: action.text
-      });
-    case FETCH_PROJECTS:
-      return _merge({}, state, {
-        projects: {
-          list: [],
-          fetching: true
-        }
-      });
-    case RECEIVE_PROJECTS:
-      return _merge({}, state, {
-        projects: {
-          list: action.response.projects,
-          fetching: false
-        }
-      });
-    case SEND_EMAIL:
-      return _merge({}, state, {
-        contact: {
-          isLoading: true
-        }
-      });
-    case RECEIVE_EMAIL:
-      if (action.response.ok) {
-        const tempState = Object.assign({}, state);
-        const fields = tempState.contact.fields;
-
-        fields.forEach((field) => {
-          field.value = '';
-          field.valid = true;
-          field.errorMsg = '';
-        });
-
-
-        return _merge({}, state, {
-          contact: {
-            isLoading: false,
-            showSuccess: true,
-            showError: false
-          }
-        });
-      }
-
-      return _merge({}, state, {
-        contact: {
-          isLoading: false,
-          showError: true
-        }
-      });
-    case DISMISS_SUCCESS:
-      return _merge({}, state, {
-        contact: {
-          showSuccess: false
-        }
-      });
-    case DISMISS_ERROR:
-      return _merge({}, state, {
-        contact: {
-          showError: false
-        }
-      });
-    case UPDATE_FIELD: {
-      const tempState = Object.assign({}, state);
-      const field = _find(tempState.contact.fields, { name: action.name });
-      field.value = action.value;
-
-      return _merge({}, state, tempState);
-    }
-    case UPDATE_FIELDS: {
-      const tempState = Object.assign({}, state);
-      const fields = tempState.contact.fields;
-      tempState.contact.fields = _merge(fields, action.fields);
-
-      return _merge({}, state, tempState);
-    }
-    default:
-      return state;
-  }
-}
-
-export default reduce;
+export default combineReducers({
+  contact,
+  pageTitle,
+  projects
+});
